[Edit] UserRepositoryTest 멤버 UserRepositoryTest 클래스로 이동
