+++ src/test/java/com/inq/wishhair/wesharewishhair/hairstyle/domain/HairStyleRepositoryTest.java
@@ -1,0 +1,6 @@
+package com.inq.wishhair.wesharewishhair.hairstyle.domain;
+
+import com.inq.wishhair.wesharewishhair.common.base.RepositoryTest;
+
+public class HairStyleRepositoryTest extends RepositoryTest {
+}
